remove-background-fastapi/requirements.txt
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
coloredlogs==15.0.1
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.103.2
flatbuffers==23.5.26
gitdb==4.0.10
GitPython==3.1.37
h11==0.14.0
humanfriendly==10.0
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.13.0
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.2
lazy_loader==0.3
llvmlite==0.41.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
nest-asyncio==1.5.8
networkx==3.1
numba==0.58.0
numpy==1.24.4
onnxruntime==1.16.0
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
pooch==1.7.0
prompt-toolkit==3.0.39
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyMatting==1.1.8
python-dateutil==2.8.2
python-multipart==0.0.6
pytz==2023.3.post1
PyWavelets==1.4.1
pyzmq==25.1.1
referencing==0.30.2
rembg==2.0.50
requests==2.31.0
rich==13.6.0
rpds-py==0.10.3
scikit-image==0.21.0
scipy==1.10.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
streamlit==1.27.1
sympy==1.12
tenacity==8.2.3
tifffile==2023.7.10
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.5
uvicorn==0.23.2
validators==0.22.0
watchdog==3.0.0
wcwidth==0.2.8
zipp==3.17.0
